Renamed consistency checker CLIs
