+++ sarkit/verification/_cphd_consistency.py
@@ -2,14 +2,12 @@
 Functionality for verifying CPHD files for internal consistency.
 """
 
-import argparse
 import collections
 import functools
 import itertools
 import logging
 import numbers
 import os
-import pathlib
 import re
 import struct
 from typing import Any, Optional
@@ -23,11 +21,6 @@ import sarkit.cphd as skcphd
 import sarkit.verification._consistency as con
 import sarkit.wgs84
 from sarkit import _constants
-
-try:
-    from smart_open import open
-except ImportError:
-    pass
 
 logger = logging.getLogger(__name__)
 
@@ -1982,42 +1975,3 @@ def calc_refgeom_parameters(xml, xmlhelp, pvps):
     return collections.namedtuple("refgeom_params", "refgeom monostat bistat")(
         refgeom, mono, bistat
     )
-
-
-def _parser():
-    parser = argparse.ArgumentParser(
-        description="Analyze a CPHD and display inconsistencies"
-    )
-    parser.add_argument("file_name", help="CPHD or CPHD XML to check")
-    parser.add_argument(
-        "--schema",
-        type=pathlib.Path,
-        help="Use a supplied schema file (attempts version-specific schema if omitted)",
-    )
-    parser.add_argument(
-        "--thorough",
-        action="store_true",
-        help=(
-            "Run checks that may seek/read through large portions of the file. "
-            "Ignored when file_name is XML"
-        ),
-    )
-    CphdConsistency.add_cli_args(parser)
-    return parser
-
-
-def main(args=None):
-    config = _parser().parse_args(args)
-    with open(config.file_name, "rb") as f:
-        cphd_con = CphdConsistency.from_file(
-            file=f,
-            schema=config.schema,
-            thorough=config.thorough,
-        )
-        return cphd_con.run_cli(config)
-
-
-if __name__ == "__main__":  # pragma: no cover
-    import sys
-
-    sys.exit(int(main()))
