+++ sarkit/verification/_crsd_consistency.py
@@ -2,11 +2,9 @@
 Functionality for verifying CRSD files for internal consistency.
 """
 
-import argparse
 import functools
 import logging
 import os
-import pathlib
 import re
 from typing import Any, Optional
 
@@ -18,11 +16,6 @@ from lxml import etree
 import sarkit.crsd as skcrsd
 import sarkit.verification._consistency as con
 import sarkit.wgs84
-
-try:
-    from smart_open import open
-except ImportError:
-    pass
 
 logger = logging.getLogger(__name__)
 
@@ -2418,42 +2411,3 @@ class CrsdConsistency(con.ConsistencyChecker):
                     assert self.xmlhelp.load_elem(xml_val_node) == con.Approx(
                         geom[key], atol=1e-6, rtol=0
                     )
-
-
-def _parser():
-    parser = argparse.ArgumentParser(
-        description="Analyze a CRSD and display inconsistencies"
-    )
-    parser.add_argument("file_name", help="CRSD or CRSD XML to check")
-    parser.add_argument(
-        "--schema",
-        type=pathlib.Path,
-        help="Use a supplied schema file (attempts version-specific schema if omitted)",
-    )
-    parser.add_argument(
-        "--thorough",
-        action="store_true",
-        help=(
-            "Run checks that may seek/read through large portions of the file. "
-            "Ignored when file_name is XML"
-        ),
-    )
-    CrsdConsistency.add_cli_args(parser)
-    return parser
-
-
-def main(args=None):
-    config = _parser().parse_args(args)
-    with open(config.file_name, "rb") as f:
-        crsd_con = CrsdConsistency.from_file(
-            file=f,
-            schema=config.schema,
-            thorough=config.thorough,
-        )
-        return crsd_con.run_cli(config)
-
-
-if __name__ == "__main__":  # pragma: no cover
-    import sys
-
-    sys.exit(int(main()))
