+++ sarkit/verification/_sicd_consistency.py
@@ -2,13 +2,11 @@
 Functionality for verifying SICD files for internal consistency.
 """
 
-import argparse
 import copy
 import datetime
 import functools
 import logging
 import os
-import pathlib
 from typing import Any, Optional
 
 import numpy as np
@@ -23,11 +21,6 @@ import sarkit.sicd.projection as sicdproj
 import sarkit.verification._consistency as con
 import sarkit.wgs84
 from sarkit import _constants
-
-try:
-    from smart_open import open
-except ImportError:
-    pass
 
 logger = logging.getLogger(__name__)
 
@@ -1964,34 +1957,3 @@ class SicdConsistency(con.ConsistencyChecker):
             with self.precondition():
                 assert poly_node is not None
                 self._assert_poly_2d(poly_node, poly)
-
-
-def _parser():
-    parser = argparse.ArgumentParser(
-        description="Analyze a SICD and display inconsistencies"
-    )
-    parser.add_argument("file_name", help="SICD or SICD XML to check")
-    parser.add_argument(
-        "--schema",
-        type=pathlib.Path,
-        help="Use a supplied schema file (attempts version-specific schema if omitted)",
-    )
-    SicdConsistency.add_cli_args(parser)
-    return parser
-
-
-def main(args=None):
-    config = _parser().parse_args(args)
-    with open(config.file_name, "rb") as f:
-        sicd_con = SicdConsistency.from_file(
-            file=f,
-            schema=config.schema,
-        )
-    # file doesn't need to stay open once object is instantiated
-    return sicd_con.run_cli(config)
-
-
-if __name__ == "__main__":  # pragma: no cover
-    import sys
-
-    sys.exit(int(main()))
