+++ tests/verification/test_cli.py
@@ -3,8 +3,6 @@ import subprocess
 import pytest
 
 
-@pytest.mark.parametrize(
-    "cmd", ("cphd-consistency", "crsd-consistency", "sicd-consistency")
-)
+@pytest.mark.parametrize("cmd", ("cphdcheck", "crsdcheck", "sicdcheck", "siddcheck"))
 def test_consistency_cli(cmd):
     subprocess.run([cmd, "-h"], check=True)
