+++ tests/verification/test_crsd_consistency.py
@@ -12,7 +12,8 @@ from lxml import etree
 
 import sarkit.crsd as skcrsd
 import sarkit.wgs84
-from sarkit.verification._crsd_consistency import CrsdConsistency, main
+from sarkit.verification._crsd_consistency import CrsdConsistency
+from sarkit.verification._crsdcheck import main
 
 from . import testing
